pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sarpyx"
version = "0.1.0"
description = "A swissknife for SAR processing."
authors = [
    { name = "Roberto Del Prete", email = "roberto.delprete@esa.int" }
]
license = { text = "MIT" }
requires-python = ">=3.11, <3.13"
dependencies = [
    "matplotlib",
    "scipy",
    "s1isp @ git+https://github.com/avalentino/s1isp.git",
    "geopandas",
    "shapely",
    "numpy",
    "openpyxl",
    "h5py",
    "numba",
    "jupyter",
    "torch>=1.9.0",
    "phidown>=0.1.12",
    "rasterio>=1.4.3",
    "zarr>=3.0.8",
    "joblib>=1.5.1",
    "dask>=2025.5.1",
    "lxml>=5.4.0",
    "huggingface-hub[cli,torch]>=0.33.0",
    "tifffile>=2025.6.11",
]

[project.scripts]
sarpyx = "sarpyx.cli.main:main"
sarpyx-shipdet = "sarpyx.cli.shipdet:main"

[tool.setuptools]
packages = ["sarpyx"]

[tool.setuptools.package-data]
"*" = ["*.csv", "*.json", "*.xml"]
[dependency-groups]
dev = [
    "pytest>=8.4.0",
]
